+++ src/main/java/org/example/model/tipoCasilleros/DeMulta.java
@@ -3,8 +3,8 @@ package org.example.model.tipoCasilleros;
 import org.example.model.Jugador;
 
 public class DeMulta extends Casillero implements CasilleroEjecutable {
-    private int montoMulta;
-    public DeMulta(int ubicacion,int montoMulta){
+    private double montoMulta;
+    public DeMulta(int ubicacion,double montoMulta){
         super("Multa",TipoCasillero.MULTA,ubicacion);
         this.montoMulta = montoMulta;
     }
